refactor(113):tornando o login compativel com a segurança do sistema

+++ src/main/java/br/com/chronos/servicologin/autenticacao/PayloadUsuario.java
@@ -3,10 +3,9 @@ package br.com.chronos.servicologin.autenticacao;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record PayloadUsuario(
-        Integer id,
+        Long id,
         String nome,
         String email,
         @JsonProperty("cargo_id")
         Long cargoId
-) {
-}
+) {}
