+++ src/main/java/br/com/chronos/servicologin/usuario/Usuario.java
@@ -1,10 +1,9 @@
 package br.com.chronos.servicologin.usuario;
 
 public record Usuario(
-        Integer id,
+        Long id,
         String nome,
         String email,
         String passwordHash,
         Long cargoId
-) {
-}
+) {}
